+++ CHANGELOG.md
@@ -1,10 +1,29 @@
 # Next
 
 Git diff:
-* https://github.com/readium/r2-streamer-js/compare/v1.0.22...develop
+* https://github.com/readium/r2-streamer-js/compare/v1.0.23...develop
 
 Changes:
 * TODO
+
+# 1.0.23
+
+> Build environment: NodeJS `12.13.0`, NPM `6.12.0`
+
+Changes:
+* NPM package updates
+* TAJSON now parses/generates arbitrary JSON properties with typed object
+* OPDS2 browser (with data:image/xxx,Base64 handling, and navigation into the v1-v2 converter / inspector)
+
+Git revision info:
+* https://unpkg.com/r2-streamer-js@1.0.23/dist/gitrev.json
+* https://github.com/edrlab/r2-streamer-js-dist/blob/v1.0.23/dist/gitrev.json
+
+Git commit history:
+* https://github.com/readium/r2-streamer-js/commits/v1.0.23
+
+Git diff:
+* https://github.com/readium/r2-streamer-js/compare/v1.0.22...v1.0.23
 
 # 1.0.22